+++ data/smbios_data.py
@@ -2808,6 +2808,7 @@ smbios_dictionary = {
         # VMware Virtual Machine
         "Board ID": "440BX Desktop Reference Platform",
         "FirmwareFeatures": None,
+        "SecureBootModel": "x86legacy",
         "CPU Generation": cpu_data.cpu_data.penryn.value,
         "Max OS Supported": os_data.os_data.max_os,
         "Wireless Model": None,
@@ -2819,6 +2820,7 @@ smbios_dictionary = {
         # Parallels Virtual Machine
         "Board ID": "Parallels Virtual Platform",
         "FirmwareFeatures": None,
+        "SecureBootModel": "x86legacy",
         "CPU Generation": cpu_data.cpu_data.penryn.value,
         "Max OS Supported": os_data.os_data.max_os,
         "Wireless Model": None,
@@ -2843,6 +2845,128 @@ smbios_dictionary = {
         "Board ID": None,
         "FirmwareFeatures": None,
         "SecureBootModel": "j214",
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_GPU_Test_Board": {
+        # AppleInternal GPU Test Board
+        # System is used for testing Intel iGPUs
+        # Model has been configured with the following:
+        # - Broadwell iGPU (HD Graphics 5300)
+        # - Skylake iGPU (HD Graphics 5xx)
+        # - Kaby Lake iGPU (HD Graphics 6xx)
+        # - Ice Lake iGPU (G4/G7)
+        "Board ID": "Mac-50619A408DB004DA",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_1": {
+        # Laptop model
+        # No mentions in SLE outside X86PlatformPlugin
+        "Board ID": "Mac-112B0A653D3AAB9C",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_2": {
+        # Laptop model
+        # Only mentioned in ApplePlatformEnabler and X86PlatformPlugin
+        "Board ID": "Mac-4BFBC784B845591E",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_3": {
+        # Laptop model, 15"/17" with dGPU and GMUX
+        # Only mentioned in AppleGraphicsPowerManagement, AppleMuxControl2, AGDCBacklightControl, ApplePlatformEnabler and X86PlatformPlugin
+        "Board ID": "Mac-564FBA6031E5946A",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+     "APPL_UNKNOWN_MODEL_4": {
+        # Laptop model, iGPU only
+        # Only mentioned in AppleGraphicsPowerManagement, ApplePlatformEnabler and X86PlatformPlugin
+        "Board ID": "Mac-6FEBD60817C77D8A",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_5": {
+        # Laptop model, iGPU only, Iris Plus Graphics G7
+        # Only mentioned in AppleGraphicsPowerManagement, ApplePlatformEnabler and X86PlatformPlugin
+        "Board ID": "Mac-747B1AEFF11738BE",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_6": {
+        # Laptop model, iGPU only
+        # Only mentioned in AppleGraphicsPowerManagement, ApplePlatformEnabler and X86PlatformPlugin
+        "Board ID": "Mac-87DCB00F4AD77EEA",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_7": {
+        # Laptop model
+        # Only mentioned in X86PlatformPlugin, AppleGVA
+        "Board ID": "Mac-90BE64C3CB5A9AEB",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
+        "CPU Generation": None,
+        "Max OS Supported": os_data.os_data.max_os,
+        "Wireless Model": None,
+        "Bluetooth Model": bluetooth_data.bluetooth_data.NonApplicable,
+        "Stock GPUs": [],
+        "Stock Storage": [],
+    },
+    "APPL_UNKNOWN_MODEL_7": {
+        # Laptop model, iGPU only
+        # Only mentioned in AppleGraphicsPowerManagement and X86PlatformPlugin
+        "Board ID": "Mac-9394BDF4BF862EE7",
+        "FirmwareFeatures": None,
+        "SecureBootModel": None,
         "CPU Generation": None,
         "Max OS Supported": os_data.os_data.max_os,
         "Wireless Model": None,